// src/API.ts
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type Favorites = {
  __typename: "Favorites",
  id: string,
  user?: string | null,
  jobID: Job,
  createdAt: string,
  updatedAt: string,
};
// ...
export type User = {
  __typename: "User",
  id: string,
  fullname?: string | null,
  reviewsLength?: number | null,
  totalJobsDone?: number | null,
  email?: string | null,
  phoneNumber?: string | null,
  avatar?: string | null,
  notificationsActived?: boolean | null,
  website?: string | null,
  firstTimeLoggin?: boolean | null,
  categories?: Array< string | null > | null,
  Locations?: ModelLocationConnection | null,
  Portifolios?: ModelPortifolioConnection | null,
  hasDeficiency?: boolean | null,
  Favorites?: ModelFavoritesConnection | null,
  ChatRooms?: ModelUserChatRoomConnection | null,
  Messages?: ModelMessageConnection | null,
  points?: number | null,
  profission?: string | null,
  backgroundImage?: string | null,
  Notifications?: ModelNotificationConnection | null,
  selectedPortifolio?: Portifolio | null,
  status?: UserStatus | null,
  jobsHistoric?: ModelJobHistoricModelConnection | null,
  ActivityHistoric?: ModelActivityHistoricConnection | null,
  about?: string | null,
  cv?: string | null,
  createdAt: string,
  updatedAt: string,
  userSelectedPortifolioId?: string | null,
};

export type ModelLocationConnection = {
  __typename: "ModelLocationConnection",
  items:  Array<Location | null >,
  nextToken?: string | null,
};

export type Location = {
  __typename: "Location",
  id: string,
  city?: string | null,
  state?: string | null,
  lat?: string | null,
  long?: string | null,
  country?: string | null,
  street?: string | null,
  userID?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelPortifolioConnection = {
  __typename: "ModelPortifolioConnection",
  items:  Array<Portifolio | null >,
  nextToken?: string | null,
};

export type Portifolio = {
  __typename: "Portifolio",
  id: string,
  title?: string | null,
  filesKey?: Array< string | null > | null,
  backgroundImageKey?: string | null,
  content?: string | null,
  userID: string,
  startedDate?: string | null,
  endDate?: string | null,
  techologiesUsed?: Array< string | null > | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelFavoritesConnection = {
  __typename: "ModelFavoritesConnection",
  items:  Array<Favorites | null >,
  nextToken?: string | null,
};

export type ModelUserChatRoomConnection = {
  __typename: "ModelUserChatRoomConnection",
  items:  Array<UserChatRoom | null >,
  nextToken?: string | null,
};

export type UserChatRoom = {
  __typename: "UserChatRoom",
  id: string,
  userID: string,
  chatRoomID: string,
  user: User,
  chatRoom: ChatRoom,
  createdAt: string,
  updatedAt: string,
};

export type ChatRoom = {
  __typename: "ChatRoom",
  id: string,
  name?: string | null,
  Messages?: ModelMessageConnection | null,
  users?: ModelUserChatRoomConnection | null,
  LastMessage?: Message | null,
  Attachments?: ModelAttachmentConnection | null,
  createdAt: string,
  updatedAt: string,
  chatRoomLastMessageId?: string | null,
};

export type ModelMessageConnection = {
  __typename: "ModelMessageConnection",
  items:  Array<Message | null >,
  nextToken?: string | null,
};

export type Message = {
  __typename: "Message",
  id: string,
  text?: string | null,
  chatroomID: string,
  userID: string,
  images?: Array< string | null > | null,
  Attachments?: ModelAttachmentConnection | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type ModelNotificationConnection = {
  __typename: "ModelNotificationConnection",
  items:  Array<Notification | null >,
  nextToken?: string | null,
};

export type Notification = {
  __typename: "Notification",
  id: string,
  text: string,
  userID: string,
  jobID: string,
  read?: boolean | null,
  senderID: string,
  notificationType?: NotificationType | null,
  createdAt: string,
  updatedAt: string,
};

export enum NotificationType {
  APPLIED = "APPLIED",
  COMPLETED = "COMPLETED",
  PAID = "PAID",
}


export enum UserStatus {
  REPORTED = "REPORTED",
  DELETED = "DELETED",
  FREQUENT = "FREQUENT",
  RARELY = "RARELY",
  NORMAL = "NORMAL",
}


export type ModelJobHistoricModelConnection = {
  __typename: "ModelJobHistoricModelConnection",
  items:  Array<JobHistoricModel | null >,
  nextToken?: string | null,
};

export type JobHistoricModel = {
  __typename: "JobHistoricModel",
  id: string,
  userID?: string | null,
  startDate?: string | null,
  endDate?: string | null,
  company?: string | null,
  jobTitle?: string | null,
  jobDescription?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelActivityHistoricConnection = {
  __typename: "ModelActivityHistoricConnection",
  items:  Array<ActivityHistoric | null >,
  nextToken?: string | null,
};

export type ActivityHistoric = {
  __typename: "ActivityHistoric",
  id: string,
  userID: string,
  type?: ActivityType | null,
  typeID?: string | null,
  typeText?: string | null,
  createdAt: string,
  updatedAt: string,
};

export enum ActivityType {
  MESSAGE = "MESSAGE",
  JOB = "JOB",
  USER = "USER",
  PORTIFOLIO = "PORTIFOLIO",
}
// ...
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type Category = {
  __typename: "Category",
  id: string,
  name: string,
  userID?: string | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type ModelUserStatusInput = {
  eq?: UserStatus | null,
  ne?: UserStatus | null,
};
// ...
export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelNotificationTypeInput = {
  eq?: NotificationType | null,
  ne?: NotificationType | null,
};
// ...
export type ModelStatusInput = {
  eq?: Status | null,
  ne?: Status | null,
};
// ...
export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export type ModelFloatInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelDifficultyEnumsInput = {
  eq?: DifficultyEnums | null,
  ne?: DifficultyEnums | null,
};
// ...
export type ModelPortifolioFilterInput = {
  id?: ModelIDInput | null,
  title?: ModelStringInput | null,
  filesKey?: ModelStringInput | null,
  backgroundImageKey?: ModelStringInput | null,
  content?: ModelStringInput | null,
  userID?: ModelIDInput | null,
  startedDate?: ModelStringInput | null,
  endDate?: ModelStringInput | null,
  techologiesUsed?: ModelStringInput | null,
  and?: Array< ModelPortifolioFilterInput | null > | null,
  or?: Array< ModelPortifolioFilterInput | null > | null,
  not?: ModelPortifolioFilterInput | null,
};
// ...
export type ModelCategoryFilterInput = {
  id?: ModelIDInput | null,
  name?: ModelStringInput | null,
  userID?: ModelIDInput | null,
  and?: Array< ModelCategoryFilterInput | null > | null,
  or?: Array< ModelCategoryFilterInput | null > | null,
  not?: ModelCategoryFilterInput | null,
};

export type ModelCategoryConnection = {
  __typename: "ModelCategoryConnection",
  items:  Array<Category | null >,
  nextToken?: string | null,
};
// ...
export type ModelNotificationFilterInput = {
  id?: ModelIDInput | null,
  text?: ModelStringInput | null,
  userID?: ModelIDInput | null,
  jobID?: ModelIDInput | null,
  read?: ModelBooleanInput | null,
  senderID?: ModelStringInput | null,
  notificationType?: ModelNotificationTypeInput | null,
  and?: Array< ModelNotificationFilterInput | null > | null,
  or?: Array< ModelNotificationFilterInput | null > | null,
  not?: ModelNotificationFilterInput | null,
};
// ...
export type ModelUserFilterInput = {
  id?: ModelIDInput | null,
  fullname?: ModelStringInput | null,
  reviewsLength?: ModelIntInput | null,
  totalJobsDone?: ModelIntInput | null,
  email?: ModelStringInput | null,
  phoneNumber?: ModelStringInput | null,
  avatar?: ModelStringInput | null,
  notificationsActived?: ModelBooleanInput | null,
  website?: ModelStringInput | null,
  firstTimeLoggin?: ModelBooleanInput | null,
  categories?: ModelStringInput | null,
  hasDeficiency?: ModelBooleanInput | null,
  points?: ModelIntInput | null,
  profission?: ModelStringInput | null,
  backgroundImage?: ModelStringInput | null,
  status?: ModelUserStatusInput | null,
  about?: ModelStringInput | null,
  cv?: ModelStringInput | null,
  and?: Array< ModelUserFilterInput | null > | null,
  or?: Array< ModelUserFilterInput | null > | null,
  not?: ModelUserFilterInput | null,
  userSelectedPortifolioId?: ModelIDInput | null,
};
// ...
export type ModelJobFilterInput = {
  id?: ModelIDInput | null,
  title?: ModelStringInput | null,
  description?: ModelStringInput | null,
  dateTimestamp?: ModelIntInput | null,
  dateDate?: ModelStringInput | null,
  location?: ModelStringInput | null,
  canBeDoneRemotely?: ModelBooleanInput | null,
  minimumPrice?: ModelFloatInput | null,
  qualifications?: ModelStringInput | null,
  difficulty?: ModelDifficultyEnumsInput | null,
  status?: ModelStatusInput | null,
  urgent?: ModelBooleanInput | null,
  needPreviousExperience?: ModelBooleanInput | null,
  categories?: ModelStringInput | null,
  points?: ModelIntInput | null,
  completedAt?: ModelStringInput | null,
  and?: Array< ModelJobFilterInput | null > | null,
  or?: Array< ModelJobFilterInput | null > | null,
  not?: ModelJobFilterInput | null,
  jobAuthorId?: ModelIDInput | null,
  jobJobDoneById?: ModelIDInput | null,
};
// ...
export type ListPortifoliosQueryVariables = {
  filter?: ModelPortifolioFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListPortifoliosQuery = {
  listPortifolios?:  {
    __typename: "ModelPortifolioConnection",
    items:  Array< {
      __typename: "Portifolio",
      id: string,
      title?: string | null,
      filesKey?: Array< string | null > | null,
      backgroundImageKey?: string | null,
      content?: string | null,
      userID: string,
      startedDate?: string | null,
      endDate?: string | null,
      techologiesUsed?: Array< string | null > | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};
// ...
export type ListCategoriesQueryVariables = {
  filter?: ModelCategoryFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListCategoriesQuery = {
  listCategories?:  {
    __typename: "ModelCategoryConnection",
    items:  Array< {
      __typename: "Category",
      id: string,
      name: string,
      userID?: string | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};
// ...
export type ListNotificationsQueryVariables = {
  filter?: ModelNotificationFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListNotificationsQuery = {
  listNotifications?:  {
    __typename: "ModelNotificationConnection",
    items:  Array< {
      __typename: "Notification",
      id: string,
      text: string,
      userID: string,
      jobID: string,
      read?: boolean | null,
      senderID: string,
      notificationType?: NotificationType | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};
// ...
export type ListUsersQueryVariables = {
  filter?: ModelUserFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};
// ...
export type ListJobsQueryVariables = {
  filter?: ModelJobFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};